# Optimize findblockindex