@fredrikekre 2022-12-07
There is https://github.com/JuliaLang/julia/blob/db00cc1a8455ceb4a8dc3cb8dd6ded3d62e46dcb/base/sort.jl#L262-L276 but none for UnitRange's. Perhaps better to add that instead of here. This was just something I found while investigating the overhad of indexing a BlockArray vs an Array.