@dlfivefifty 2022-12-08
We could add it here temporarily, but make it clearer, say as a function call ` _searchsortedfirst ` with a comment that it can be removed in the future?