@fredrikekre 2022-12-08
I updated the PR with another approach of computing the block and index simultaneously. This cuts the time for the benchmark in half.